third_party/blink/web_tests/virtual/lna-allow-potentially-trustworthy-same-origin/external/wpt/fetch/private-network-access/xhr-from-treat-as-public.https.window-expected.txt

This is a testharness.js-based test.
FAIL treat-as-public to local: failed preflight. assert_equals: response loaded expected false but got true
PASS treat-as-public to local: success.
PASS treat-as-public to private: failed preflight.
PASS treat-as-public to private: success.
PASS treat-as-public to public: no preflight required.
Harness: the test ran to completion.
